refactor(ocrypto): split KeyType and ECCMode out of ec_key_pair.go
Move the cross-cutting KeyType algorithm identifier (and its parse,
predicate, and conversion helpers) into key_type.go, and the ECCMode
enum (with its constants, String, ECSizeToMode, and
GetECCurveFromECCMode) into ecc_mode.go. Pure same-package relocation
with no logic or call-site changes; ec_key_pair.go now focuses on EC/
ML-KEM key pairs.
