+++ lib/ocrypto/ec_key_pair.go
@@ -18,49 +18,6 @@ import (
 	"golang.org/x/crypto/hkdf"
 )
 
-type ECCMode uint8
-
-type KeyType string
-
-const (
-	RSA2048Key   KeyType = "rsa:2048"
-	RSA4096Key   KeyType = "rsa:4096"
-	EC256Key     KeyType = "ec:secp256r1"
-	EC384Key     KeyType = "ec:secp384r1"
-	EC521Key     KeyType = "ec:secp521r1"
-	MLKEM768Key  KeyType = "mlkem:768"
-	MLKEM1024Key KeyType = "mlkem:1024"
-)
-
-// ParseKeyType validates a string as a known KeyType, returning an error for
-// unrecognized values.
-func ParseKeyType(alg string) (KeyType, error) {
-	switch KeyType(alg) {
-	case RSA2048Key, RSA4096Key,
-		EC256Key, EC384Key, EC521Key,
-		MLKEM768Key, MLKEM1024Key,
-		HybridXWingKey, HybridSecp256r1MLKEM768Key, HybridSecp384r1MLKEM1024Key:
-		return KeyType(alg), nil
-	default:
-		return "", fmt.Errorf("unrecognized key type: %s", alg)
-	}
-}
-
-const (
-	ECCModeSecp256r1 ECCMode = 0
-	ECCModeSecp384r1 ECCMode = 1
-	ECCModeSecp521r1 ECCMode = 2
-	ECCModeSecp256k1 ECCMode = 3
-)
-
-const (
-	ECCurveP256Size = 256
-	ECCurveP384Size = 384
-	ECCurveP521Size = 521
-	RSA2048Size     = 2048
-	RSA4096Size     = 4096
-)
-
 type KeyPair interface {
 	PublicKeyInPemFormat() (string, error)
 	PrivateKeyInPemFormat() (string, error)
@@ -102,106 +59,6 @@ type MLKEMKeyPair struct {
 
 type MLKEM1024KeyPair struct {
 	PrivateKey *mlkem.DecapsulationKey1024
-}
-
-func IsECKeyType(kt KeyType) bool {
-	switch kt { //nolint:exhaustive // only handle ec types
-	case EC256Key, EC384Key, EC521Key:
-		return true
-	default:
-		return false
-	}
-}
-
-func IsRSAKeyType(kt KeyType) bool {
-	switch kt { //nolint:exhaustive // only handle rsa types
-	case RSA2048Key, RSA4096Key:
-		return true
-	default:
-		return false
-	}
-}
-
-func IsMLKEMKeyType(kt KeyType) bool {
-	switch kt { //nolint:exhaustive // only handle mlkem types
-	case MLKEM768Key, MLKEM1024Key:
-		return true
-	default:
-		return false
-	}
-}
-
-// GetECCurveFromECCMode return elliptic curve from ecc mode
-func GetECCurveFromECCMode(mode ECCMode) (elliptic.Curve, error) {
-	var c elliptic.Curve
-
-	switch mode {
-	case ECCModeSecp256r1:
-		c = elliptic.P256()
-	case ECCModeSecp384r1:
-		c = elliptic.P384()
-	case ECCModeSecp521r1:
-		c = elliptic.P521()
-	case ECCModeSecp256k1:
-		// TODO FIXME - unsupported?
-		return nil, errors.New("unsupported ECC mode")
-	default:
-		return nil, fmt.Errorf("unsupported ECC mode %d", mode)
-	}
-
-	return c, nil
-}
-
-func (mode ECCMode) String() string {
-	switch mode {
-	case ECCModeSecp256r1:
-		return "ec:secp256r1"
-	case ECCModeSecp384r1:
-		return "ec:secp384r1"
-	case ECCModeSecp521r1:
-		return "ec:secp521r1"
-	case ECCModeSecp256k1:
-		return "ec:secp256k1"
-	}
-	return "unspecified"
-}
-
-// ECSizeToMode converts a curve size to an ECCMode
-func ECSizeToMode(size int) (ECCMode, error) {
-	switch size {
-	case ECCurveP256Size:
-		return ECCModeSecp256r1, nil
-	case ECCurveP384Size:
-		return ECCModeSecp384r1, nil
-	case ECCurveP521Size:
-		return ECCModeSecp521r1, nil
-	default:
-		return 0, fmt.Errorf("unsupported EC curve size: %d", size)
-	}
-}
-
-func ECKeyTypeToMode(kt KeyType) (ECCMode, error) {
-	switch kt { //nolint:exhaustive // only handle ec types
-	case EC256Key:
-		return ECCModeSecp256r1, nil
-	case EC384Key:
-		return ECCModeSecp384r1, nil
-	case EC521Key:
-		return ECCModeSecp521r1, nil
-	default:
-		return 0, fmt.Errorf("unsupported type: %v", kt)
-	}
-}
-
-func RSAKeyTypeToBits(kt KeyType) (int, error) {
-	switch kt { //nolint:exhaustive // only handle rsa types
-	case RSA2048Key:
-		return RSA2048Size, nil
-	case RSA4096Key:
-		return RSA4096Size, nil
-	default:
-		return 0, fmt.Errorf("unsupported type: %v", kt)
-	}
 }
 
 // NewECKeyPair Generates an EC key pair of the given bit size.
